+++ Tools/generalTools.cpp
