+++ Tools/generalTools.h
@@ -1,8 +1,8 @@
 #ifndef GENERALTOOLS_H
 #define GENERALTOOLS_H
-#include "Tile.h"
-#include "Tileset.h"
-#include "Board.h"
+#include "../Classes/Board.h"
+#include "../Classes/Tileset.h"
+#include "../Classes/Board.h"
 
 Tileset edgeTiles(Board* toTest, bool goingRight);
 int getRandomNumber(int low, int high);
